+++ vvp_app_server/settings.py
@@ -90,6 +90,7 @@ INSTALLED_APPS = [
     "notifications",
     "factApi",
     "reportFake",
+    "contact",
     "payment",
     "proxycache",
     "info",